requirements.txt

BareNecessities==0.2.8
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Mail==0.9.1
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dotenv==0.15.0
redis==3.5.3
requests==2.24.0
urllib3==1.25.11
Werkzeug==1.0.1
